model: bugfix - mo_may_allow was too restrictive
This bugfix has the same rationale as with the following commit:

commit 4ecb2a814963776ea80db47ae4b29b3c8baba867

We do not want the potential reader to disqualify itself from reading a
future value; previously, this was fixed only for reads (not RMW's). But
we can experience the same problem with RMW's, which are both read and
write.

So, instead of using 'act != reader' as a special case for the is_read()
case, just use it as a breaking condition in addition to

   !(reader --hb-> act)

This is really intended anyway, since our happens-before is reflexive,
whereas it is irreflexive in the specification.